+++ test/helper.rb
@@ -10,7 +10,9 @@ require 'fluent/plugin/buf_memory'
 require 'fluent/plugin/buf_file'
 require 'fluent/test/driver/output'
 
-require 'fluent/plugin/out_bigquery'
+require 'fluent/plugin/out_bigquery_base'
+require 'fluent/plugin/out_bigquery_insert'
+require 'fluent/plugin/out_bigquery_load'
 require 'google/apis/bigquery_v2'
 require 'google/api_client/auth/key_utils'
 require 'googleauth'
